haciendo todo otra vez pero con firebase

+++ app/src/main/java/com/naldana/ejemplo10/CurrencyAdder.kt
@@ -23,7 +23,7 @@ class CurrencyAdder : AppCompatActivity() {
     private fun gater():Coin = Coin(
         findViewById<EditText>(R.id.editText_name).text.toString(),
         findViewById<EditText>(R.id.editText_country).text.toString(),
-        findViewById<EditText>(R.id.editText_year).text.toString(),
+        findViewById<EditText>(R.id.editText_year).text.toString().toLong(),
         findViewById<CheckBox>(R.id.available).isChecked
     )
 }
